// frontend/src/types/sports/Court.ts
import type { SportFilter } from "../sports/SportFilter";

export interface Court {
  id: number;
  nombre: string;
  deporte: SportFilter;
  descripcion: string;
  precio: number;
  precioFinde: number;
  capacidad: string;
  superficie: string;
  techada: boolean;
  imagen: string;
  disponible: boolean;
  destacada: boolean;
  amenidades: string[];
}